Refactoring assemble command.
