+++ tests/test_assemble.py
@@ -21,18 +21,20 @@ import os
 from pathlib import Path
 
 from proksee.reads import Reads
-from proksee.commands.cmd_assemble import assemble
+from proksee.pipelines.assemble import assemble
 from proksee.resource_specification import ResourceSpecification
 
 INPUT_DIR = os.path.join(Path(__file__).parent.absolute(), "data")
 OUTPUT_DIR = os.path.join(Path(__file__).parent.absolute(), "output")
 RESOURCE_SPECIFICATION = ResourceSpecification(4, 4)  # 4 threads, 4 gigabytes
 
+DATABASE_PATH = os.path.join(Path(__file__).parent.parent.absolute(), "proksee", "database",
+                             "refseq_short.csv")
 TEST_MASH_DB_FILENAME = os.path.join(Path(__file__).parent.absolute(), "data", "ecoli.msh")
 TEST_ID_MAPPING_FILENAME = os.path.join(Path(__file__).parent.absolute(), "data", "test_id_mapping.tab")
 
 
-class TestCmdAssemble:
+class TestAssemble:
 
     def test_simple_assemble(self):
         """
@@ -67,9 +69,8 @@ class TestCmdAssemble:
             os.remove(json_file)
 
         assemble(reads, OUTPUT_DIR, force,
-                 TEST_MASH_DB_FILENAME, RESOURCE_SPECIFICATION,
-                 species_name=None, platform_name=None,
-                 id_mapping_filename=TEST_ID_MAPPING_FILENAME)
+                 DATABASE_PATH, TEST_MASH_DB_FILENAME, RESOURCE_SPECIFICATION,
+                 TEST_ID_MAPPING_FILENAME, species_name=None, platform_name=None)
 
         # Check that expected files were created:
         assert os.path.isfile(assembly_statistics_file)
